Introduce FileParser
