--- src/Parser/AbstractCfonbParser.php
@@ -1,22 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the CFONB Parser package.
- *
- * (c) Guillaume Sainthillier <hello@silarhi.fr>
- * (c) @fezfez <demonchaux.stephane@gmail.com>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
-namespace Silarhi\Cfonb\Parser;
-
-use Silarhi\Cfonb\Contracts\ParserInterface;
-
-abstract class AbstractCfonbParser implements ParserInterface
-{
-    abstract protected function getSupportedCode(): string;
-}
